requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
pyyaml>=6.0
pytest>=7.0.0
anthropic>=0.40.0
